% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{align_test_sets}
\alias{align_test_sets}
\title{Align RF test set with LSTM test set}
\usage{
align_test_sets(
  test_dataset,
  lstm_test_dict,
  ts_index_info,
  site_name_col,
  datetime_col = "time"
)
}
\arguments{
\item{test_dataset}{Original test data.frame}

\item{lstm_test_dict}{LSTM test dictionary (with datetime element)}

\item{ts_index_info}{Index info from lstm_specific_preprocessing}

\item{site_name_col}{Site name column}

\item{datetime_col}{Datetime column}
}
\value{
Filtered test data.frame aligned to LSTM endpoints
}
\description{
Filters the point-based test dataset to only include rows that were
successfully processed as window endpoints by the LSTM windowing.
}
\examples{
data(microclimate_sample)
splits <- split_train_val_test(microclimate_sample, train_pct = 0.6, val_pct = 0.2, block_days = 2, seed = 42)
scaled <- lstm_scaling(splits$train, splits$val, splits$test)
prep <- lstm_specific_preprocessing(scaled$train, scaled$val, scaled$test,
                                    window_size = 2, ts_names_col = "time_series_doc")
aligned <- align_test_sets(splits$test, prep$test_dict, prep$index_info, "time_series_doc")
nrow(aligned)
}

