% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.default_cols}
\alias{.default_cols}
\title{Default column names used throughout the package}
\format{
An object of class \code{list} of length 6.
}
\usage{
.default_cols
}
\description{
Default column names used throughout the package
}
\keyword{internal}
