% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_cyclical_time}
\alias{add_cyclical_time}
\title{Add cyclical time features}
\usage{
add_cyclical_time(df, datetime_col = "time", add_month = FALSE)
}
\arguments{
\item{df}{A data.frame}

\item{datetime_col}{Name of the datetime column}

\item{add_month}{Logical. Whether to also add month cyclical features.}
}
\value{
The data.frame with added Hour_sin, Hour_cos (and optionally Month_sin, Month_cos).
}
\description{
Adds sine/cosine encoding of Hour (and optionally Month) to a data.frame.
}
