% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_correction_model}
\alias{load_correction_model}
\title{Load a correction model from disk}
\usage{
load_correction_model(path)
}
\arguments{
\item{path}{File path to the .rds model}
}
\value{
List with model, scaler, feature_cols, model_type
}
\description{
Load a correction model from disk
}
