% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{make_windows}
\alias{make_windows}
\title{Create sliding windows for LSTM input}
\usage{
make_windows(
  X_mat,
  y_vec,
  base_pred_vec,
  datetime_vec,
  window_size,
  max_gap_hours = 1
)
}
\arguments{
\item{X_mat}{Matrix of features (n_samples x n_features)}

\item{y_vec}{Vector of target values}

\item{base_pred_vec}{Vector of base predictions}

\item{datetime_vec}{Vector of POSIXct datetimes}

\item{window_size}{Integer window length}

\item{max_gap_hours}{Maximum allowed gap in hours between consecutive points. NULL to disable.}
}
\value{
List with X (3D array), y, base_pred, datetime
}
\description{
Rearranges time-series data into sliding windows of fixed size.
Skips windows that contain temporal gaps larger than max_gap_hours.
}
\examples{
\dontrun{
win <- make_windows(
  X_mat         = as.matrix(site_a_train[, feat_cols]),
  y_vec         = site_a_train$residual,
  base_pred_vec = site_a_train$predicted,
  datetime_vec  = site_a_train$time,
  window_size   = 6,
  max_gap_hours = 1
)
}
}
