% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_correction_model}
\alias{save_correction_model}
\title{Save a correction model to disk}
\usage{
save_correction_model(model, scaler, feature_cols, path)
}
\arguments{
\item{model}{Trained model (ranger or keras)}

\item{scaler}{List with min/max from scaling}

\item{feature_cols}{Character vector of feature column names}

\item{path}{File path to save to (will create .rds file)}
}
\description{
Save a correction model to disk
}
\examples{
data(microclimate_sample)
feature_cols <- c("TAREF", "RH", "VREF", "SOLR")
rf <- train_rf(microclimate_sample[, feature_cols], microclimate_sample$residual, num_trees = 5, tune = FALSE)
tmp <- tempfile(fileext = ".rds")
save_correction_model(rf, scaler = NULL, feature_cols = feature_cols, path = tmp)
unlink(tmp)
}

